Created a background service that collect user location passively
The service only run if the user used any another app that has access to user location eg; google maps.
The service is optimized and does not consume too much battery because it only run if there is a gps data requested by any other app and it does not run if the change in the location is less than 1 kilo meter.